Replace all ImageButton examples with just one
